[Visibility] Export Elasticsearch new client util functions
